+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add Sublime Text syntax highlighting.
 - Add function/process/entity and name getters to `Unit`.
+- Add `--dump` option to `llhd-check`.
 
 ### Changed
 - Fix missing `dyn` keywords.